# feat: optimize plant state updates with lazy Redis caching & unify API logic